moved getting taskList from TaskPage to App
